requirements.txt
apache-airflow==2.8.1
pandas
boto3
python-dotenv